refactor pets page
