+++ app/src/main/java/com/example/cs446/ui/pages/main/pets/EditPetDialog.kt
@@ -1,0 +1,107 @@
+package com.example.cs446.ui.pages.main.pets
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.cs446.data.model.Pet
+
+@Composable
+fun EditPetDialog(
+    pet: Pet,
+    onDismiss: () -> Unit,
+    onSave: (String, String, String) -> Unit
+) {
+    var name by remember { mutableStateOf(pet.name) }
+    var breed by remember { mutableStateOf(pet.breed ?: "") }
+    var weight by remember { mutableStateOf(pet.weight.toString()) }
+
+    var nameError by remember { mutableStateOf(false) }
+    var weightError by remember { mutableStateOf(false) }
+
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        confirmButton = {
+            TextButton(onClick = {
+                // Validation
+                nameError = name.isBlank()
+                weightError = weight.toDoubleOrNull() == null || weight.toDoubleOrNull()!! <= 0.0
+
+                if (!nameError && !weightError) {
+                    onSave(name, breed, weight)
+                }
+            }) {
+                Text("Save")
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDismiss) {
+                Text("Cancel")
+            }
+        },
+        title = { Text("Edit Pet") },
+        text = {
+            Column {
+                OutlinedTextField(
+                    value = name,
+                    onValueChange = {
+                        name = it
+                        nameError = false
+                    },
+                    label = { Text("Pet Name") },
+                    isError = nameError,
+                    modifier = Modifier.fillMaxWidth()
+                )
+                if (nameError) {
+                    Text(
+                        "Name cannot be empty.",
+                        color = MaterialTheme.colorScheme.error,
+                        fontSize = 12.sp
+                    )
+                }
+
+                Spacer(modifier = Modifier.height(8.dp))
+
+                OutlinedTextField(
+                    value = breed,
+                    onValueChange = { breed = it },
+                    label = { Text("Breed") },
+                    modifier = Modifier.fillMaxWidth()
+                )
+
+                Spacer(modifier = Modifier.height(8.dp))
+
+                OutlinedTextField(
+                    value = weight,
+                    onValueChange = {
+                        weight = it
+                        weightError = false
+                    },
+                    label = { Text("Weight (lbs)") },
+                    isError = weightError,
+                    modifier = Modifier.fillMaxWidth()
+                )
+                if (weightError) {
+                    Text(
+                        "Weight must be a positive number.",
+                        color = MaterialTheme.colorScheme.error,
+                        fontSize = 12.sp
+                    )
+                }
+            }
+        }
+    )
+}
